pyproject.toml
[tool.poetry]
name = "trustyai-tests"
version = "0.1.0"
description = ""
authors = ["aaguirre <aaguirre@redhat.com>"]
readme = "README.md"
packages = [{include = "trustyai_tests"}]

[tool.poetry.dependencies]
python = "^3.11"
kubernetes = "^29.0.0"
openshift-python-wrapper = "^10.0.0"
pytest = "^8.2.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
